normalize image if not

+++ gwtd/utils/standardization.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-IMAGE_MEAN = np.array([0.485, 0.456, 0.406])
-IMAGE_STD = np.array([0.229, 0.224, 0.225])
+IMAGE_MEAN = np.array([0.485, 0.456, 0.406], dtype=np.float32)
+IMAGE_STD = np.array([0.229, 0.224, 0.225], dtype=np.float32)
 
 def standardize_image(image):
     """
